services/server/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
psycopg2==2.8.5
python-dotenv==0.13.0